fix: move all commands to flexible, removing old createcommand
