+++ packages/tools/kadena-cli/src/devnet/commands/devnetList.ts
@@ -1,15 +1,13 @@
-import type { CreateCommandReturnType } from '../../utils/createCommand.js';
 import { createCommand } from '../../utils/createCommand.js';
 import { log } from '../../utils/logger.js';
 import { displayDevnetsConfig } from '../utils/devnetDisplay.js';
 
-export const listDevnetsCommand: CreateCommandReturnType = createCommand(
+export const listDevnetsCommand = createCommand(
   'list',
   'List all available devnets',
   [],
-  async (config) => {
-    log.debug('devnet-list:action', { config });
-
+  async () => {
+    log.debug('devnet-list:action');
     await displayDevnetsConfig();
   },
 );
