+++ packages/tools/kadena-cli/src/devnet/commands/devnetSimulation.ts
@@ -1,4 +1,3 @@
-import type { CreateCommandReturnType } from '../../utils/createCommand.js';
 import { createCommand } from '../../utils/createCommand.js';
 import { globalOptions } from '../../utils/globalOptions.js';
 import { log } from '../../utils/logger.js';
@@ -6,7 +5,7 @@ import { networkIsAlive } from '../utils/network.js';
 import { simulateCoin } from '../utils/simulation/coin/simulate.js';
 import { simulationOptions } from '../utils/simulation/simulationOptions.js';
 
-export const simulateCommand: CreateCommandReturnType = createCommand(
+export const simulateCommand = createCommand(
   'simulate',
   'Simulate traffic on a devnet',
   [
@@ -20,8 +19,9 @@ export const simulateCommand: CreateCommandReturnType = createCommand(
     simulationOptions.simulationSeed({ isOptional: true }),
     simulationOptions.simulationMaxTime({ isOptional: true }),
   ],
-  async (config) => {
-    log.debug('devnet-simulate:action', { config });
+  async (option, { collect }) => {
+    const config = await collect(option);
+    log.debug('devnet-simulate:action', config);
 
     if (!(await networkIsAlive(config.networkConfig.networkHost))) {
       log.info(
