+++ packages/tools/kadena-cli/src/networks/commands/networkList.ts
@@ -1,16 +1,11 @@
-import { displayNetworksConfig } from '../utils/networkDisplay.js';
-
 import type { Command } from 'commander';
+
 import { createCommand } from '../../utils/createCommand.js';
 import { log } from '../../utils/logger.js';
+import { displayNetworksConfig } from '../utils/networkDisplay.js';
 
 export const listNetworksCommand: (program: Command, version: string) => void =
-  createCommand(
-    'list',
-    'List all available local networks',
-    [],
-    async (config) => {
-      log.debug('network-list:action', { config });
-      await displayNetworksConfig();
-    },
-  );
+  createCommand('list', 'List all available local networks', [], async () => {
+    log.debug('network-list:action');
+    await displayNetworksConfig();
+  });
