+++ packages/tools/kadena-cli/src/tx/index.ts
@@ -1,5 +1,3 @@
-import { createCommandFlexible } from '../utils/createCommandFlexible.js';
-import { globalOptions } from '../utils/globalOptions.js';
 import { createTransactionCommandNew } from './commands/txCreateTransaction.js';
 import { createSendTransactionCommand } from './commands/txSend.js';
 import { createSignCommand } from './commands/txSign.js';
@@ -8,19 +6,6 @@ import { createTestSignedTransactionCommand } from './commands/txTestSignedTrans
 import type { Command } from 'commander';
 
 const SUBCOMMAND_ROOT: 'tx' = 'tx';
-
-// execute with: pnpm run dev tx testing --account-name="123" --public-keys="123,123"
-export const testCommand: (program: Command, version: string) => void =
-  createCommandFlexible(
-    'testing',
-    '',
-    [globalOptions.accountName(), globalOptions.publicKeys()],
-    async (option, { collect }) => {
-      // TODO: runtime value is correct, but type is not. Objects need to be merged in the types
-      const config = await collect(option);
-      console.log(config);
-    },
-  );
 
 export function txCommandFactory(program: Command, version: string): void {
   const txProgram = program
@@ -31,5 +16,4 @@ export function txCommandFactory(program: Command, version: string): void {
   createSignCommand(txProgram, version);
   createTestSignedTransactionCommand(txProgram, version);
   createTransactionCommandNew(txProgram, version);
-  testCommand(txProgram, version);
 }
