+++ packages/tools/kadena-cli/src/tx/utils/txSignWithLocalWallet.ts
@@ -8,7 +8,7 @@ import type { EncryptedString } from '@kadena/hd-wallet';
 import type { IWallet } from '../../keys/utils/keysHelpers.js';
 import { getWalletContent } from '../../keys/utils/keysHelpers.js';
 
-import type { CommandOption } from '../../utils/createCommandFlexible.js';
+import type { CommandOption } from '../../utils/createCommand.js';
 import { log } from '../../utils/logger.js';
 import type { options } from '../commands/txSignOptions.js';
 import { parseTransactionsFromStdin } from './input.js';
