--- packages/tools/kadena-cli/src/utils/createCommandFlexible.ts
@@ -1,226 +1,0 @@
-import type { Command } from 'commander';
-import { CommandError, printCommandError } from './command.util.js';
-import { getCommandExecution } from './createCommand.js';
-import type { OptionType } from './createOption.js';
-import { globalOptions } from './globalOptions.js';
-import { handlePromptError } from './helpers.js';
-import { log } from './logger.js';
-import { readStdin } from './stdin.js';
-import type { Fn, Prettify } from './typeUtilities.js';
-
-export type OptionConfig<Option extends OptionType> = {
-  [P in Option['key']]: Option['transform'] extends Fn
-    ? Awaited<ReturnType<Option['transform']>>
-    : Awaited<ReturnType<Option['prompt']>>;
-} & (Option['expand'] extends Fn
-  ? {
-      [P in `${Option['key']}Config`]: Awaited<ReturnType<Option['expand']>>;
-    }
-  : {});
-
-type PromptFn = (
-  args: Record<string, unknown>,
-  originalArgs: Record<string, unknown>,
-) => unknown;
-
-export async function executeOption<Option extends OptionType>(
-  option: Option,
-  args: Record<string, unknown> = {},
-  originalArgs: Record<string, unknown> = {},
-): Promise<
-  Prettify<{
-    value: unknown;
-    config: OptionConfig<Option>;
-  }>
-> {
-  let value = args[option.key];
-
-  if (value === undefined && option.isInQuestions) {
-    if (args.quiet !== true && args.quiet !== 'true') {
-      value = await (option.prompt as PromptFn)(args, originalArgs);
-    } else if (option.isOptional === false) {
-      throw new Error(
-        `Missing required argument: ${option.key} (${option.option.flags})`,
-      );
-    }
-  }
-
-  const validate = (
-    option.isOptional ? option.validation.optional() : option.validation
-  ).safeParse(value);
-  if (validate.success === false) {
-    log.warning(`Invalid value for ${option.key}: ${value}`);
-  }
-
-  const newConfig = { [option.key]: value } as OptionConfig<Option>;
-  if ('expand' in option) {
-    if (typeof option.expand === 'function') {
-      const expanded = await option.expand(value, args);
-      if (expanded !== undefined && expanded !== null) {
-        // @ts-ignore
-        newConfig[`${option.key}Config`] = expanded;
-      }
-    }
-  }
-  if ('transform' in option) {
-    if (typeof option.transform === 'function') {
-      // @ts-ignore
-      newConfig[option.key] = await option.transform(value, args);
-    }
-  }
-
-  return {
-    value: value,
-    config: newConfig,
-  };
-}
-
-const printCommandExecution = (
-  command: string,
-  args: Record<string, unknown>,
-  updateArgs?: Record<string, unknown>,
-): void => {
-  // Give the option to update args used in the command by returning an object
-  // Only update args that are already defined
-  if (updateArgs !== undefined && typeof updateArgs === 'object') {
-    for (const [key, value] of Object.entries(updateArgs)) {
-      if (Object.hasOwn(args, key) === true) args[key] = value;
-    }
-  }
-
-  log.info(
-    log.color.yellow(`\nExecuted: ${getCommandExecution(command, args)}`),
-  );
-};
-
-export type CommandOption<T extends OptionType[]> = {
-  [K in T[number]['key']]: (
-    customArgs?: Record<string, unknown>,
-  ) => Promise<Prettify<OptionConfig<Extract<T[number], { key: K }>>>>;
-};
-
-interface ICommandData {
-  /** Arguments parsed from the command line */
-  values: string[];
-  /** content passed to stdin */
-  stdin?: string;
-  /** Automatically fetch all options in order of the options array and merge the results */
-  collect: <T extends CommandOption<OptionType[]>>(
-    options: T,
-  ) => Promise<{
-    [K in keyof T]: Awaited<ReturnType<T[K]>>;
-  }>;
-}
-
-export const createCommandFlexible =
-  <
-    T extends OptionType[],
-    C extends (
-      option: CommandOption<T>,
-      data: ICommandData,
-    ) => Promise<Record<string, unknown> | void>,
-  >(
-    name: string,
-    description: string,
-    options: T,
-    action: C,
-  ): ((program: Command, version: string) => void) =>
-  (program) => {
-    // anything after the first newline is only shown on the command specific help page
-    const [_description, ...helpText] = description.split('\n');
-    let command = program.command(name).description(_description);
-    if (helpText.length > 0) {
-      command.configureHelp({
-        commandDescription: () =>
-          `${[_description, '', ...helpText].join('\n')}`,
-      });
-    }
-    let allowsUnknownOptions = false;
-
-    if (options.some((option) => option.allowUnknownOptions === true)) {
-      command = command.allowUnknownOption(true);
-      allowsUnknownOptions = true;
-    }
-
-    command.addOption(globalOptions.quiet().option);
-    options.forEach((option) => {
-      command.addOption(option.option);
-    });
-    command.action(async (originalArgs, ...rest) => {
-      // args outside try-catch to be able to use it in catch
-      let args = { ...originalArgs };
-
-      try {
-        const stdin = await readStdin();
-
-        if (allowsUnknownOptions) {
-          log.debug(`Command ${name} allows unknown options`);
-        }
-
-        // Automatically enable quiet mode if not in interactive environment
-        if (!process.stderr.isTTY) args.quiet = true;
-
-        const optionIndex = new Map<unknown, number>();
-        const collectOptionsMap = options.reduce((acc, option, index) => {
-          acc[option.key] = async (customArgs = {}) => {
-            try {
-              const { value, config } = await executeOption(
-                option,
-                {
-                  ...args,
-                  ...customArgs,
-                },
-                originalArgs,
-              );
-
-              // Keep track of previous args to prompts can use them
-              args = { ...args, [option.key]: value };
-              return config;
-            } catch (error) {
-              handlePromptError(error);
-            }
-          };
-          optionIndex.set(acc[option.key], index);
-          return acc;
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        }, {} as any);
-
-        const values = rest.flatMap((r) => r.args);
-        const result = await action(collectOptionsMap, {
-          values,
-          stdin: stdin ?? undefined,
-          collect: async (optionObject) => {
-            const options = Object.entries(optionObject);
-            options.sort(
-              (a, b) =>
-                (optionIndex.get(a[1]) ?? 0) - (optionIndex.get(b[1]) ?? 0),
-            );
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            let result = {} as any;
-            for (const option of options) {
-              result = {
-                ...result,
-                ...(await option[1]()),
-              };
-            }
-            return result;
-          },
-        });
-
-        printCommandExecution(
-          `${program.name()} ${name}`,
-          args,
-          result ?? undefined,
-        );
-      } catch (error) {
-        if (error instanceof CommandError) {
-          printCommandError(error);
-          printCommandExecution(`${program.name()} ${name}`, args, error.args);
-          process.exitCode = error.exitCode;
-          return;
-        }
-        log.error(`\nAn error occurred: ${error.message}\n`);
-        process.exitCode = 1;
-      }
-    });
-  };
